chore: api 및 hook 기능별로 한 파일에  정리
