+++ src/api/diary/useWriteDiary.ts
@@ -1,0 +1,32 @@
+import { apiRoutes } from "@api/apiRoutes";
+import api from "@api/fetcher";
+import { useMutation } from "@tanstack/react-query";
+import { useNavigate } from "react-router-dom";
+import { WriteDiaryPayLoad, WriteDiaryResponse } from "src/types/diaryTypes";
+
+// 일기 작성
+const postDiary = async (diaryData: WriteDiaryPayLoad) => {
+  try {
+    const { data }: { data: WriteDiaryResponse } = await api.post({
+      endpoint: apiRoutes.diary,
+      body: diaryData,
+    });
+    return data;
+  } catch (error) {
+    console.log(error);
+    throw error;
+  }
+};
+
+export const useWriteDiary = () => {
+  const navigate = useNavigate();
+  return useMutation({
+    mutationFn: (diaryData: WriteDiaryPayLoad) => postDiary(diaryData),
+    onSuccess: (data) => {
+      navigate(`/diary/${data.id}`);
+    },
+    onError: (error) => {
+      console.log(error);
+    },
+  });
+};
