+++ src/types/diaryTypes.ts
@@ -1,3 +1,5 @@
+import { DiaryFormData } from "src/types/WriteDiaryTypes";
+
 export type BaseDiaryType = {
   id: number;
   date: string;
@@ -9,12 +11,11 @@ export type DiaryListType = BaseDiaryType & {
   title: string;
 };
 
-export type DiaryDataType = BaseDiaryType & {
+export type WriteDiaryPayLoad = {
+  date: string;
   nickname: string;
-  count?: number;
-  isFull?: boolean;
-  mood: string;
-  weather: string;
-  title: string;
-  story: string;
+} & DiaryFormData;
+
+export type WriteDiaryResponse = {
+  id: number;
 };
